requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
google-generativeai==0.8.6
chromadb>=0.5.0
pypdf2==3.0.1
python-dotenv==1.0.0
